+++ tests/test_providers/test_numbers.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 import re
+import decimal
 
 import pytest
 
@@ -36,7 +37,7 @@ class TestNumbers(object):
         result = numbers.floats(n=1000)
         assert len(result) == 1000
 
-        result = numbers.floats(rounding=4)
+        result = numbers.floats(precision=4)
         for e in result:
             assert len(str(e).split('.')[1]) <= 4
 
@@ -57,6 +58,16 @@ class TestNumbers(object):
         element = numbers.random.choice(result)
         assert isinstance(element, int)
 
+    def test_decimals(self, numbers, start, end):
+        result = numbers.decimals(start=start, end=end)
+
+        assert max(result) <= end
+        assert min(result) >= start
+        assert isinstance(result, list)
+
+        element = numbers.random.choice(result)
+        assert isinstance(element, decimal.Decimal)
+
     @pytest.mark.parametrize(
         'start_real, end_real, start_imag, end_imag', [
             (1.2, 10, 1, 2.4),
@@ -65,9 +76,9 @@ class TestNumbers(object):
         ],
     )
     def test_complexes(self, numbers,
-                         start_real, end_real, start_imag, end_imag):
+                       start_real, end_real, start_imag, end_imag):
         result = numbers.complexes(start_real, end_real,
-                                     start_imag, end_imag)
+                                   start_imag, end_imag)
         assert max(e.real for e in result) <= end_real
         assert min(e.real for e in result) >= start_real
         assert max(e.imag for e in result) <= end_imag
@@ -78,7 +89,7 @@ class TestNumbers(object):
         result = numbers.complexes(n=1000)
         assert len(result) == 1000
 
-        result = numbers.complexes(rounding_real=4, rounding_imag=6)
+        result = numbers.complexes(precision_real=4, precision_imag=6)
         for e in result:
             assert len(str(e.real).split('.')[1]) <= 4
             assert len(str(e.imag).split('.')[1]) <= 6
@@ -88,7 +99,7 @@ class TestNumbers(object):
         with pytest.raises(NonEnumerableError):
             numbers.matrix(num_type='int')
 
-        result = numbers.matrix(rounding=4)
+        result = numbers.matrix(precision=4)
         assert len(result) == 10
         for row in result:
             assert len(row) == 10
@@ -105,7 +116,7 @@ class TestNumbers(object):
                 assert isinstance(e, int)
 
         result = numbers.matrix(
-            num_type=NumTypes.COMPLEXES, rounding_real=4, rounding_imag=6)
+            num_type=NumTypes.COMPLEXES, precision_real=4, precision_imag=6)
         assert len(result) == 10
         for row in result:
             assert len(row) == 10
@@ -113,35 +124,22 @@ class TestNumbers(object):
                 assert len(str(e.real).split('.')[1]) <= 4
                 assert len(str(e.imag).split('.')[1]) <= 6
 
-    def test_primes(self, numbers):
-        result = numbers.primes()
-        assert len(result) == 168
-        assert isinstance(result, list)
-
-        result = numbers.primes(500, 500000)
-        assert len(result) == 41443
-        assert isinstance(result, list)
-
-        result = numbers.primes(start=10, end=200)[1]
-        assert result == 13
-
-    def test_digit(self, numbers):
-        digits = (
-            0, 1, 2,
-            3, 4, 5,
-            6, 7, 8,
-            9,
-        )
-        result = numbers.digit()
-        assert result in digits
-
-        result = numbers.digit(to_bin=True)
-        assert isinstance(result, str)
-
-    def test_between(self, numbers):
-        result = numbers.between(90, 100)
-        assert result >= 90
+    def test_integer(self, numbers):
+        result = numbers.integer(-100, 100)
+        assert isinstance(result, int)
+        assert result >= -100
         assert result <= 100
+
+    def test_float(self, numbers):
+        result = numbers.float(-100, 100, precision=15)
+        assert isinstance(result, float)
+        assert result >= -100
+        assert result <= 100
+        assert len(str(result).split('.')[1]) <= 15
+
+    def test_decimal(self, numbers):
+        result = numbers.decimal(-100, 100)
+        assert isinstance(result, decimal.Decimal)
 
 
 class TestSeededNumbers(object):
@@ -161,7 +159,7 @@ class TestSeededNumbers(object):
     def test_integers(self, n1, n2):
         assert n1.integers() == n2.integers()
         assert n1.integers(start=-999, end=999, n=10) == \
-            n2.integers(start=-999, end=999, n=10)
+               n2.integers(start=-999, end=999, n=10)
 
     def test_complexes(self, n1, n2):
         assert n1.complexes() == n2.complexes()
@@ -171,15 +169,11 @@ class TestSeededNumbers(object):
         assert n1.matrix() == n2.matrix()
         assert n1.matrix(n=5) == n2.matrix(n=5)
 
-    def test_digit(self, n1, n2):
-        assert n1.digit() == n2.digit()
-        assert n1.digit(to_bin=True) == n2.digit(to_bin=True)
+    def test_integer(self, n1, n2):
+        assert n1.integer() == n2.integer()
 
-    def test_between(self, n1, n2):
-        assert n1.between() == n2.between()
-        assert n1.between(minimum=42, maximum=2048) == \
-            n2.between(minimum=42, maximum=2048)
+    def test_float(self, n1, n2):
+        assert n1.float() == n2.float()
 
-    @pytest.mark.skip(reason='Method refactoring needed.')
-    def test_primes(self, n1, n2):
-        assert n1.primes() == n2.primes()
+    def test_decimal(self, n1, n2):
+        assert n1.decimal() == n2.decimal()